Close #326: CLANG warnings about some enum hackery suppressed
It's a pretty convoluted workaround for some annoying C++ restrictions.
(And some CLANG versions -- unfortunately including the one in the GitHub
Ubuntu GHA runner... -- even crashed on it, so it used to be even more
convoluted because of having to work around that, too. This commit finally
removes at least that part, as CLANG seems to have healed by now.)