Close #347: Null-safe widget lookup + access API
Not a complete solution yet, just a basic WidgetPtr<WidgetType> "smart"
pointer wrapper for the results of widget lookups by name.

Also: this is not the originally conceived "dummy widget" approach (which
is not feasible yet, or at all), but just an extra indirection over the
legacy raw Widget pointers (which have to go anyway sooner or later anyway;
but note: this change has not yet eliminated all of those either!).

* Build: GCC -std=c++23 (for std::expected<>), MSVC -Zc:preprocessor