Start #392: First shot at an "SFML shield" adapter layer
With the recently frozen pre-3 SFML API the existing (but reorganized)
low-level gfx. elements now compile, plus the layouts and the overall
GUI manager, but the actual "widgeting" doesn't, as most of that stuff
is still just raw SFML client code.

But that will change, and an adapter layer is being grown to support that.

(Checking this in, in this half-baked state, for backup, basically.
Yeah, well, can't be bothered with git branching: this lib is legacy
code anyway (marching toward starting afresh in a new repo), mostly just
a testbed now for a more intresting adapter lib design, coming up...)