First shot at #436: Wrap away the SFML Event API
It's already much nicer; the event loops in the test/example apps are
almost back to SFML2 simplicity and flexibility. ;)

Now event polling can (should) also go through the GUI manager (not just
directly using the SFML API to poll the host window), and, as a bonus,
it can now filter out the massive redundant flood of raw mouse-move events
transparently. (Which the SFML guys will obviously disable at some point
in the future anyway ;) but it's nice to have it dealt with already.)